Add -receiveFrame:length:
This works like -receiveFrame:, except returns how much data
it read.

And then do some crudeness: The old -receiveFrame: would fail
if the 2nd byte of frame was 0x02. While this may be valid
for the dock protocol, this doesn't seem to be valid for the
hammer protocol.  HammerConnection is only using -receiveFrame:length:,
so relocate this 0x02 check into -receiveFrame: (which only
PackageConnection uses)