Add HammerController, an object that can (barely) speak the Hammer protocol.  Enough to read memory off the device.
With this it'll handshake to Debugger/Newtsbug Connection on
the device (aka InitSerialDebugging).

It'll try to probe to figure out what device it's running on
(via reads to gROMManufacturer, gHardwareType, gROMVersion)

It'll try to probe to figure out the ROM size, by checking 1MB
increments to see if they repeat the word found at 0x00000000
(e.g. due to ROM mirroring, this should happen).

Then it'll attempt to download the ROM at a speed that is best
described as "glacial".  Currently it's reading 12 bytes per
request.  Given an 8MB ROM...

And afterwards, you'll have a hung Newton that'll need to be
power cycled. Sniffing traffic between a Newton and Hammer
shows that the "Go" command sends 0x10 0x10 0x01, but this
doesn't seem to work here...