Store machine type / manufacturer / rom version as instance varaibles.  The ROM version is the most important here, as we need to alter commands sent to the device debugger depending on the version.
e.g. ReadPhysMem reboots NewtOS v1.0 and v1.1 devices. So
we need to always use ReadMem on these.

We don't need to do anything on v1.x to un-stop the device,
so the (not-working) sendGoCommand isn't necessary for these.

And for v1.x, we can ask to read the entire ROM region in
one call, as opposed to the 12-byte chunks we're doing with v2.x