Add argument "readResult" to -sendReadMemoryCommandWithAddress:length:
Keep existing -sendReadMemoryCommandWithAddress:length: which
invokes this with YES to readResult.

When dumping v1.x memory, we can ask for huge regions (e.g. 4MB, 8MB).
However we don't want to wait for -readDataForResponse:length: to
return all of this data – we can't update the progress indicator
in the UI, etc.

So the dumping routine for v1.x will invoke -sendReadMemoryCommandWithAddress:length:readResult:
for 0, 4/8MB, NO.  Then in the dumper's while loop, it'll invoke
-readResponseOfLength: to read in a chunk, update the UI, etc.

Additionally we need to alter which read command we use depending on
the OS version.  ReadPhysMem reboots v1.0 and v1.1 devices, so we'll
start out using ReadMem until we're certain of the device type
we're talking to.