Update ROM dumping method to support 1.x and 2.x devices.
1.x devices, when we issue a Read(Phys)Mem, we can give it
the entire 4MB/8MB range w/o issue.  So we'll do this, and
then read the results back in 128 byte chunks.

This means the first chunk will have a 5f (ResponseResult)
marker we need to skip, but all the subsequent chunks will
not.

2.x devices, when we issue a ReadPhysMem command with a
large count, we get back some data, but then it starts
sending back 5f0000 (ResponseResult, null bytes).  So we
continue to do things conservatively and only ask for 12
byte chunks.

This means all the chunks will have a 5f (ResponseResult)
marker we need to skip.