A window controller object that is displayed when "Download ROM" is clicked.
It has a tab view with directions for Newton OS v1.x and v2.x

The directions are a combination of the screencast and TOC
view, giving a textual description of the steps necessary to
get ROM dumping going -- while simultaneously providing some
screenshots of the Newton OS, pointing at the current step.

This also includes a progress panel for the downloads.
In addition to a progress indicator and textual desc (e.g. 2MB of 4MB)
it lists the detected hardware info of the Newton device
(manufacturer, hardware type, os version, rom size)