Add -installPackages:runAsModal:
This behaves as -installPackages: did (and -installPackages:
just invokes this with runAsModal:NO)

If runAsModal is NO, then the package install panel will be
shown as a sheet in the main window.

If runAsModal is YES, then the package install panel will
be ran as a modal window.

This is to help facilitate installing packages from
the DownloadWindowController -- we'd like it's sheet to
remain present.  So we'll run the package install progress
as a modal.